+++ Sources/VComponents/Components/Buttons/Rectangular Button/VRectangularButton.swift
@@ -23,6 +23,7 @@ import VCore
 public struct VRectangularButton<CustomLabel>: View where CustomLabel: View {
     // MARK: Properties
     private let uiModel: VRectangularButtonUIModel
+    @Environment(\.displayScale) private var displayScale: CGFloat
 
     @Environment(\.isEnabled) private var isEnabled: Bool
     private func internalState(_ baseButtonState: SwiftUIBaseButtonState) -> VRectangularButtonInternalState {
@@ -158,9 +159,11 @@ public struct VRectangularButton<CustomLabel>: View where CustomLabel: View {
     private func borderView(
         internalState: VRectangularButtonInternalState
     ) -> some View {
-        if uiModel.borderWidth > 0 {
+        let borderWidth: CGFloat = uiModel.borderWidth.toPoints(scale: displayScale)
+
+        if borderWidth > 0 {
             RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: borderWidth)
                 .scaleEffect(internalState == .pressed ? uiModel.backgroundPressedScale : 1)
         }
     }