+++ Sources/VComponents/Components/Buttons/Stretched Button (Loading)/VLoadingStretchedButton.swift
@@ -28,6 +28,8 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
     // MARK: Properties - UI Model
     private let uiModel: VLoadingStretchedButtonUIModel
 
+    @Environment(\.displayScale) private var displayScale: CGFloat
+
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
     private let isLoading: Bool
@@ -230,9 +232,11 @@ public struct VLoadingStretchedButton<CustomLabel>: View where CustomLabel: View
     private func borderView(
         internalState: VLoadingStretchedButtonInternalState
     ) -> some View {
-        if uiModel.borderWidth > 0 {
+        let borderWidth: CGFloat = uiModel.borderWidth.toPoints(scale: displayScale)
+
+        if borderWidth > 0 {
             RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: borderWidth)
                 .scaleEffect(internalState == .pressed ? uiModel.backgroundPressedScale : 1)
         }
     }