+++ Sources/VComponents/Components/Inputs/TextView/VTextView.swift
@@ -72,6 +72,7 @@ import VCore
 public struct VTextView: View {
     // MARK: Properties - UI Model
     private let uiModel: VTextViewUIModel
+    @Environment(\.displayScale) private var displayScale: CGFloat
 
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
@@ -186,9 +187,11 @@ public struct VTextView: View {
 
     @ViewBuilder 
     private var backgroundBorderView: some View {
-        if uiModel.borderWidth > 0 {
+        let borderWidth: CGFloat = uiModel.borderWidth.toPoints(scale: displayScale)
+
+        if borderWidth > 0 {
             RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: borderWidth)
         }
     }