+++ Sources/VComponents/Components/Inputs/TextView/VTextViewUIModel.swift
@@ -53,10 +53,10 @@ public struct VTextViewUIModel {
     )
 
     // MARK: Properties - Border
-    /// Border width. Set to `0`.
+    /// Border width. Set to `0` points.
     ///
     /// To hide border, set to `0`.
-    public var borderWidth: CGFloat = 0
+    public var borderWidth: PointPixelMeasurement = .points(0)
 
     /// Border colors.
     public var borderColors: StateColors = .clearColors
@@ -183,7 +183,7 @@ extension VTextViewUIModel {
     public static var success: Self {
         var uiModel: Self = .init()
 
-        uiModel.borderWidth = 1.5
+        uiModel.borderWidth = PointPixelMeasurement.points(1.5)
         uiModel.applySuccessColorScheme()
 
         return uiModel
@@ -193,7 +193,7 @@ extension VTextViewUIModel {
     public static var warning: Self {
         var uiModel: Self = .init()
 
-        uiModel.borderWidth = 1.5
+        uiModel.borderWidth = PointPixelMeasurement.points(1.5)
         uiModel.applyWarningColorScheme()
 
         return uiModel
@@ -203,7 +203,7 @@ extension VTextViewUIModel {
     public static var error: Self {
         var uiModel: Self = .init()
 
-        uiModel.borderWidth = 1.5
+        uiModel.borderWidth = PointPixelMeasurement.points(1.5)
         uiModel.applyErrorColorScheme()
 
         return uiModel