+++ Sources/VComponents/Components/Notifications/Notification/VNotificationUIModel.swift
@@ -49,10 +49,10 @@ public struct VNotificationUIModel {
     }()
 
     // MARK: Properties - Border
-    /// Border width. Set to `0`.
+    /// Border width. Set to `0` points.
     ///
     /// To hide border, set to `0`.
-    public var borderWidth: CGFloat = 0
+    public var borderWidth: PointPixelMeasurement = .points(0)
 
     /// Border color.
     public var borderColor: Color = .clear