+++ Sources/VComponents/Components/Notifications/Toast/VToastUIModel.swift
@@ -43,10 +43,10 @@ public struct VToastUIModel {
     public var backgroundColor: Color = .dynamic(Color(235, 235, 235), Color(60, 60, 60))
 
     // MARK: Properties - Border
-    /// Border width. Set to `0`.
+    /// Border width. Set to `0` points.
     ///
     /// To hide border, set to `0`.
-    public var borderWidth: CGFloat = 0
+    public var borderWidth: PointPixelMeasurement = .points(0)
 
     /// Border color.
     public var borderColor: Color = .clear