+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBox.swift
@@ -38,6 +38,8 @@ public struct VCheckBox<CustomLabel>: View where CustomLabel: View {
     // MARK: Properties - UI Model
     private let uiModel: VCheckBoxUIModel
 
+    @Environment(\.displayScale) private var displayScale: CGFloat
+
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
     @Binding private var state: VCheckBoxState
@@ -120,14 +122,16 @@ public struct VCheckBox<CustomLabel>: View where CustomLabel: View {
     }
     
     private var checkBoxView: some View {
-        baseButtonView(label: { internalState in
+        let borderWidth: CGFloat = uiModel.borderWidth.toPoints(scale: displayScale)
+
+        return baseButtonView(label: { internalState in
             ZStack(content: {
                 RoundedRectangle(cornerRadius: uiModel.cornerRadius)
                     .foregroundStyle(uiModel.fillColors.value(for: internalState))
 
-                if uiModel.borderWidth > 0 {
+                if borderWidth > 0 {
                     RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                        .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                        .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: borderWidth)
                 }
 
                 if let checkmarkIcon: Image = checkmarkIcon(internalState: internalState) {