+++ Sources/VComponents/Components/State Pickers/Toggle Button (Stretched)/VStretchedToggleButton.swift
@@ -54,6 +54,8 @@ public struct VStretchedToggleButton<CustomLabel>: View where CustomLabel: View 
     // MARK: Properties - UI Model
     private let uiModel: VStretchedToggleButtonUIModel
 
+    @Environment(\.displayScale) private var displayScale: CGFloat
+
     // MARK: Properties - State
     @Environment(\.isEnabled) private var isEnabled: Bool
     @Binding private var state: VRectangularToggleButtonState
@@ -223,9 +225,11 @@ public struct VStretchedToggleButton<CustomLabel>: View where CustomLabel: View 
     private func borderView(
         internalState: VStretchedToggleButtonInternalState
     ) -> some View {
-        if uiModel.borderWidth > 0 {
+        let borderWidth: CGFloat = uiModel.borderWidth.toPoints(scale: displayScale)
+
+        if borderWidth > 0 {
             RoundedRectangle(cornerRadius: uiModel.cornerRadius)
-                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: uiModel.borderWidth)
+                .strokeBorder(uiModel.borderColors.value(for: internalState), lineWidth: borderWidth)
                 .scaleEffect(internalState.isPressedOffPressedOn ? uiModel.backgroundPressedScale : 1)
         }
     }